feat: added filters, added project directory ignore all .txt files
